etl/pyproject.toml
[project]
name = "hospital_neo4j_etl"
version = "0.1"
dependencies = [
   "neo4j==5.14.1",
   "retry==0.9.2",
   "python-dotenv"
]

[project.optional-dependencies]
dev = ["black", "flake8"]